requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
anyio==3.6.2
    # via starlette
click==8.1.3
    # via
    #   nltk
    #   uvicorn
fastapi==0.89.1
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
idna==3.4
    # via anyio
joblib==1.2.0
    # via
    #   nltk
    #   scikit-learn
nltk==3.8.1
    # via -r requirements.in
numpy==1.24.2
    # via
    #   pandas
    #   scikit-learn
    #   scipy
pandas==1.5.3
    # via -r requirements.in
pydantic==1.10.4
    # via fastapi
python-dateutil==2.8.2
    # via pandas
pytz==2022.7.1
    # via pandas
pyyaml==6.0
    # via -r requirements.in
regex==2022.10.31
    # via nltk
scikit-learn==1.2.1
    # via -r requirements.in
scipy==1.10.0
    # via scikit-learn
six==1.16.0
    # via python-dateutil
sniffio==1.3.0
    # via anyio
starlette==0.22.0
    # via fastapi
threadpoolctl==3.1.0
    # via scikit-learn
tqdm==4.64.1
    # via
    #   -r requirements.in
    #   nltk
typing-extensions==4.4.0
    # via pydantic
uvicorn==0.20.0
    # via -r requirements.in
